[pulsar] feat: Add greetd/regreet & cage for greeter
This moves hyprland from home-manager into nixos.  That allows regreet
to find hyprland as a possible session and cleans up a lot of extra
configuration I was doing to enable hyprland to run properly.  This is a
barebones greeter configuration, cage does not like multiple monitors.
Sway is probably a better option here.